PhysX ids are 64 bits, not 32...

+++ COLLADAMaya/include/COLLADAMayaPhysXXML.h
@@ -65,7 +65,7 @@ namespace COLLADAMaya
         {
             Id(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int id;
+            uint64_t id;
         };
 
         struct Point
@@ -269,7 +269,7 @@ namespace COLLADAMaya
         {
             ConvexMesh(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int convexMesh;
+            uint64_t convexMesh;
         };
 
         struct PxConvexMeshGeometry
@@ -297,7 +297,7 @@ namespace COLLADAMaya
         {
             TriangleMesh(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int triangleMesh;
+            uint64_t triangleMesh;
         };
 
         struct PxTriangleMeshGeometry
@@ -363,7 +363,7 @@ namespace COLLADAMaya
         {
             PxMaterialRef(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int materialRef;
+            uint64_t materialRef;
         };
 
         struct Materials
@@ -579,14 +579,14 @@ namespace COLLADAMaya
         {
             Actor0(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int actor0;
+            uint64_t actor0;
         };
 
         struct Actor1
         {
             Actor1(xmlNode* node);
             void exportElement(COLLADASW::StreamWriter& sw);
-            int actor1;
+            uint64_t actor1;
         };
 
         struct Actors
@@ -1000,9 +1000,9 @@ namespace COLLADAMaya
             PhysX30Collection physX30Collection;
             bool validate();
 
-            PxConvexMesh* findConvexMesh(int id);
-            PxTriangleMesh* findTriangleMesh(int id);
-            PxMaterial* findMaterial(int ref);
+            PxConvexMesh* findConvexMesh(uint64_t id);
+            PxTriangleMesh* findTriangleMesh(uint64_t id);
+            PxMaterial* findMaterial(uint64_t ref);
             PxMaterial* findMaterial(const String& bodyName, const String& shapeName);
             PxShape* findShape(const String& bodyName, const String& shapeName);
             PxRigidStatic* findRigidStatic(const String& bodyName);
